+++ src/main/java/sts_exporter/CreatureExportData.java
@@ -3,6 +3,7 @@ package sts_exporter;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.HashMap;
 import java.util.HashSet;
 
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
@@ -14,27 +15,28 @@ import com.megacrit.cardcrawl.core.CardCrawlGame;
 import com.megacrit.cardcrawl.core.Settings;
 import com.megacrit.cardcrawl.daily.DailyMods;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
-import com.megacrit.cardcrawl.dungeons.Exordium;
-import com.megacrit.cardcrawl.dungeons.TheBeyond;
-import com.megacrit.cardcrawl.dungeons.TheCity;
-import com.megacrit.cardcrawl.helpers.EnemyData;
 import com.megacrit.cardcrawl.helpers.MonsterHelper;
 import com.megacrit.cardcrawl.map.MapRoomNode;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
 import com.megacrit.cardcrawl.monsters.MonsterGroup;
+import com.megacrit.cardcrawl.monsters.city.Byrd;
 import com.megacrit.cardcrawl.monsters.exordium.Lagavulin;
 import com.megacrit.cardcrawl.rooms.EmptyRoom;
 
 import basemod.BaseMod;
+import basemod.ReflectionHacks;
 
 public class CreatureExportData {
     public AbstractCreature creature;
-    public String image, relImage;
+    public String image, absImage, relImage;
     public String name;
+    public ModExportData mod;
 
     public CreatureExportData(AbstractCreature creature, String imageDir) {
         this.creature = creature;
         this.name = creature.name;
+        this.mod = Exporter.findMod(creature.getClass());
+        this.mod.creatures.add(this);
         exportImageToDir(imageDir);
     }
 
@@ -45,9 +47,10 @@ public class CreatureExportData {
         }
         safename = safename.replace(" ","");
         safename = safename.replace("/","");
-        this.image = imageDir + "/" + safename + ".png";
-        this.relImage = "creatures/" + safename + ".png";
-        exportImageToFile(this.image);
+        this.image = safename + ".png";
+        this.absImage = imageDir + "/" + this.image;
+        this.relImage = "creatures/" + this.image;
+        exportImageToFile(this.absImage);
     }
 
     private void exportImageToFile(String imageFile) {
@@ -77,9 +80,42 @@ public class CreatureExportData {
         });
     }
 
+    public static ArrayList<CreatureExportData> exportAllCreatures(String outdir) {
+        Exporter.mkdir(outdir);
+        ArrayList<CreatureExportData> creatures = new ArrayList<>();
+        for (AbstractCreature m : getAllCreatures()) {
+            creatures.add(new CreatureExportData(m, outdir));
+        }
+        return creatures;
+    }
+
     public static ArrayList<AbstractCreature> getAllCreatures() {
+        ArrayList<AbstractCreature> creatures = new ArrayList<>();
+        creatures.addAll(getAllPlayers());
+        creatures.addAll(getAllMonsters());
+        return creatures;
+    }
+
+    // Get all player characters
+    public static ArrayList<AbstractPlayer> getAllPlayers() {
         // We need to initialize DailyMods before creating AbstractPlayers
         DailyMods.setModsFalse();
+        ArrayList<AbstractPlayer> players = new ArrayList<>();
+        try {
+            Method createCharacter = CardCrawlGame.class.getDeclaredMethod("createCharacter", AbstractPlayer.PlayerClass.class);
+            createCharacter.setAccessible(true);
+            for (AbstractPlayer.PlayerClass playerClass : AbstractPlayer.PlayerClass.values()) {
+                AbstractPlayer p = (AbstractPlayer)createCharacter.invoke(null, playerClass);
+                p.name = p.title;
+                players.add(p);
+            }
+        } catch (Exception e) {
+            Exporter.logger.error("Exception occured when getting createCharacter method", e);
+        }
+        return players;
+    }
+
+    public static ArrayList<AbstractMonster> getAllMonsters() {
         // We need to initialize the random seeds before creating AbstractMonsters (for AbstractDungeon.monsterHpRng among others)
         Settings.seed = new Long(12345);
         AbstractDungeon.generateSeeds();
@@ -91,25 +127,11 @@ public class CreatureExportData {
         AbstractDungeon.currMapNode.room = new EmptyRoom();
         AbstractDungeon.currMapNode.room.monsters = new MonsterGroup(new AbstractMonster[0]); // needed to render monsters
 
-        // Get all player characters
-        ArrayList<AbstractCreature> creatures = new ArrayList<>();
-        try {
-            Method createCharacter = CardCrawlGame.class.getDeclaredMethod("createCharacter", AbstractPlayer.PlayerClass.class);
-            createCharacter.setAccessible(true);
-            for (AbstractPlayer.PlayerClass playerClass : AbstractPlayer.PlayerClass.values()) {
-                if (playerClass.toString() == "CROWBOT") continue; // old version of the game
-                AbstractPlayer p = (AbstractPlayer)createCharacter.invoke(null, playerClass);
-                p.name = p.title;
-                creatures.add(p);
-            }
-        } catch (Exception e) {
-            Exporter.logger.error("Exception occured when getting createCharacter method", e);
-        }
-
         // Now get all monsters
         // There is unfortunately no list of all monsters in the game. The best we can do is to use MonsterHelper.getEncounter
+        ArrayList<AbstractMonster> creatures = new ArrayList<>();
         HashSet<String> seenMonsters = new HashSet<>();
-        for (String encounter : getAllEncounterNames()) {
+        for (String encounter : BaseMod.encounterList) {
             Exporter.logger.info("Getting monsters for encounter " + encounter);
             MonsterGroup monsters = MonsterHelper.getEncounter(encounter);
             for (AbstractMonster monster : monsters.monsters) {
@@ -119,17 +141,34 @@ public class CreatureExportData {
                 seenMonsters.add(id);
             }
         }
+
+        // Custom monsters (BaseMod)
+        HashMap<String,BaseMod.GetMonsterGroup> customMonsters =
+            (HashMap<String,BaseMod.GetMonsterGroup>) ReflectionHacks.getPrivateStatic(BaseMod.class, "customMonsters");
+        for (BaseMod.GetMonsterGroup group : customMonsters.values()) {
+            MonsterGroup monsters = group.get();
+            for (AbstractMonster monster : monsters.monsters) {
+                String id = monster.getClass().getName();
+                if (seenMonsters.contains(id)) continue;
+                creatures.add(monster);
+                seenMonsters.add(id);
+            }
+        }
+
         // Awake lagavulin looks different
         Lagavulin lagavulin = new Lagavulin(false);
         lagavulin.id = lagavulin.id + "Awake";
         lagavulin.name = lagavulin.name + " (Awake)";
         creatures.add(lagavulin);
 
-        return creatures;
-    }
+        // Downed byrd looks different
+        Byrd byrd = new Byrd(500,500);
+        byrd.changeState(Byrd.GROUND_STATE);
+        byrd.id = byrd.id + "Grounded";
+        byrd.name = byrd.name + " (Grounded)";
+        creatures.add(byrd);
 
-    public static Collection<String> getAllEncounterNames() {
-        return BaseMod.encounterList;
+        return creatures;
     }
 
 }