+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 0.3.0
+* Use BaseMod.customMonsters to also export custom monsters
+* Export relics
+* Export potions
+* Make a per mod page of items from that mod
+
 ### 0.2.2
 * Use BaseMod.encounterList to also export monsters from custom encounters.